refactor: Extract session log cache from writers.
